IBM Cloud Providers code cleanup (part 1)
Several changes were made in the System Z and Power PC cloud provider
files to increase readability and decrease newcomer ramp-up time:
  * Functions and variables were renamed for clarity
  * Comments were added per [Go's guide](https://tip.golang.org/doc/comment)
  * Error messsages were prefaced with detail for clarity
  * Non-cloud interface functions were moved to separate files
