+++ pkg/ibm/ibmz_test.go
@@ -31,7 +31,7 @@ var _ = DescribeTable("IBMZProvider unit test",
 			"dynamic." + arch + ".private-ip":     testConfig["private-ip"],
 			"dynamic." + arch + ".disk":           testConfig["disk"],
 		}
-		provider := IBMZProvider(arch, config, systemNamespace)
+		provider := CreateIbmZConfig(arch, config, systemNamespace)
 		Expect(provider).ToNot(BeNil())
 		providerConfig := provider.(IBMZDynamicConfig)
 		Expect(providerConfig).ToNot(BeNil())
@@ -40,7 +40,6 @@ var _ = DescribeTable("IBMZProvider unit test",
 		Expect(providerConfig.Key).To(Equal("test-key"))
 		Expect(providerConfig.Subnet).To(Equal("test-subnet"))
 		Expect(providerConfig.Vpc).To(Equal("test-vpc"))
-		Expect(providerConfig.SecurityGroup).To(Equal("test-security-group"))
 		Expect(providerConfig.ImageId).To(Equal("test-image-id"))
 		Expect(providerConfig.Secret).To(Equal("test-secret"))
 		Expect(providerConfig.Url).To(Equal("test-url"))
